fixup adding default_actions to action
if you specify a default_action we don't inherit from the superclass and
initialize to an empty array if we have to (otherwise if we inherited
from the superclass and later specified actions, we wouldn't get the
correct behavior which overrides the superclass, and behavior would
shift based on order of actions and default_action in the resource
definition).

also uses set union to eliminate dups in the (allowed) actions and
eliminates dups in both cases.