Move `ohai` helper into platform_helpers, fix perf issue.
The ohai method is more relevant to platform detection in tests, so it
is moved into platform_helpers. It is also modified to use the
`OHAI_SYSTEM` constant we set in spec_helper. The prior implementation
did not correctly memoize the method, which resulted in several hundred
invocations of the ohai platform plugin during test loading, which took
~70s on my machine (before any tests ran).