Prevent indentation from being set to negative numbers
If we don't defensively program here, the result is terrible.  All that
gets raised is "negative argument" from deep within the formatter and
from then on the formatter is broken.  It will continue to raise errors
from events called from inside the Chef::Client exception handler and
you're gonna have a bad time.