Make solo tarballs work on windows; Fixes #1515
Adds proper path seperator support to the tarball path, and converts the
legacy call to tar via run_command to the Windows-tastic ShellOut. Adds
convenience functions for ShellOut helpers being called directly, a-la
run_command.

Tests are incomplete, since getting the system to behave as if it was
on windows means mocking out significant path functionality for dubious
results.