CHEF-5022: Differentiate between Windows service startup_types
A Windows service startup_type can be automatic, manual, or disabled.
This adds a startup_type attribute to specify automatic or manual, and defaults to automatic
because we currently have no functionality to set a service to manual.

This is important because if you stop a service that is automatic, Windows will change
the startup_type to manual.