Remove cookbook as argument to Metadata's constructor
Now that a cookbook's name must be specified in the metadata, metadata
is always created before the cookbook version object, so it should not
accept a CookbookVersion in the constructor.

The remaining arguments to the constructor are also removed. A survey of
this codebase and a search of GitHub shows that these arguments were
never used anywhere, and they were quite odd.

Additionally, the default license field is set to "all rights reserved",
which is a safe value. The filler values of maintainer and
maintainer_email have been removed, and now default to nil. This will
make it easier to validate that the user has set these values.