remove JSON gem usage
should fully convert to using ffi-yajl

there are still issues with JSON gem monkeypatching interacting with
chef-zero and the spec tests so we keep the requires here for the
json gem and the ffi_yajl/json_gem here.  when ohai and chef-zero are
fixed, we an just require ffi_yajl.