Fixing case of a nil recipe file with chef apply
Raising the error 'No recipe file was provided' with a fatal!
call when the recipe_name is nil. This case occurs when
chef-apply is called without a recipe name argument.

Also renamed the apply spec file to apply_spec so it is picked
up by an rspec test run. Fixed a typo in the file where it was
creating a 'Chef::Application::Recipe' for @app instead of
'Chef::Application::Apply'