guard conditionals inherit options set in execute resource
When defining any of the `user`, `group`, `environment`, `cwd` or `timeout`
options in your execute resource - or any resource inheriting from this
resource, like `script`, `bash`, `csh`, `perl`, `python`, `ruby` - they will be
inherited in the `only_if` or `not_if` conditional guards.

These values can still be overwritten if required, by appending them to the
conditional guards as a second argument, which functions the same as before this
change.