Modified env resource to break values up by delimiter before comparing.
For example, if someone were to do before this patch,
```ruby
windows_path 'C:\foo;C:\bar' do
  action :create
  ...
end
```
the path would grow each time this recipe was executed. This is because
compare_path would break up the current path with ;, but would not do
the same for the value being added.