drop ruby 1.9.x support for Chef-12
Ruby 1.9.3 is in security-only maintenance phase which means that
bugs for the core ruby interpreter will not be fixed and Chef cannot
support ruby 1.9.x.  Users should be using omnibus-chef packages
which come with a built-in ruby 2.x or should upgrade their own
installations to ruby 2.x (and accept that ongoing upgrade and
support cost).  Chef 11 will continue to be shipped on ruby 1.9.x
until Chef 11 is EOL.