Add Chef::Org model class
Chef Server 12 added multi-tenancy to the open source Chef Server. An
organization is the Chef Servers name for a tenant.  The Chef::Org
class provides a model of an organization to facilitate writting knife
commands that can interact with (create, destroy, list) Chef
organizations.

This implementation is copied directly from knife-opc, the knife
plugin currently in use by chef-server-ctl:

  https://github.com/opscode/knife-opc