Lock version of database cookbook
A release of mysql-chef_gem broke the database cookbook because
it removed the default recipe which the database cookbook was using.
Since the database cookbook did not lock versions of mysql-chef_gem,
it pulls the latest mysql-chef_gem cookbook and fails.

An update to the database cookbook was released to fix this, however
Berks dep solving is getting the latest mysql-chef_gem and an older database
cookbook for the webapp test cookbook.