fix apt default_release attribute broken in 12.0
the apt provider attempts to check if it has a provider that
responds to :default_release (i.e. a real apt_package resource) via
if the provider attribute is explicitly set or not, which is relying
on a side-effect which broke in 12.0.  the correct behavior is just
to duck-type the new_resource object and see if it
respond_to?(:default_release) and then use it if it does.

Also the unit tests were not testing the apt_provider against both the
package and apt_package resources.  Added some hacky spec code (which
needs a cleanup) to at least exercise the use cases.