deep_merge_cache fixes for bugs in 12.0.0
In 12.0.0 we introduced a cache for the merged attributes for the
top-level node attribute keys.  This fixes this so that node['foo']
and node[:foo] are not cached separately.  This also showed up in bugs
as issues between node['foo'] access and node.foo access because
node.foo is translated into node[:foo].