make include_recipe "::foo" use current cookbook
The following code is brittle if you want to fork the cookbook and rename it
(which is a best practice when forking cookbooks for internal use):

```ruby
if node[:platform_family] == "rhel"
  include_recipe "mycookbook::_rhel"
end
```

In order for cookbooks to be easily renamable they can currently use the
syntax `include_recipe "#{cookbook_name}::_rhel"` which is unwieldy.
This patch adds `include_recipe "::_rhel"` as syntax sugar to make this
easier.