fix master
broken by merging #2431.  code was actually broken and the specs
were broken, not sure how it got into ready-to-merge in that state.

rolled back the FileUtils.rm_rf that was in the original patch since
it trashed my chef git repo and in light of:

https://github.com/ValveSoftware/steam-for-linux/issues/3671

i think the rm_rf is a bad idea.