Remove non-default defaults from mdadm resource
The purpose of this commit is to start a conversation about the current
default values for the mdadm resource.  Which currently chooses some default
values for optional parameters.  It seems odd for Chef to choose defaults for
optional parameters where the underlaying OS/application has it's own
defaults.

As a result this commit does not modify or extend the rspec tests to cover
these changes.  I'll address this in a future commit pending the discussion.

This commit removes the default values for the following optional parameters
when creating an mdadm device.

* chunk size: Chef 16k, mdadm 512k
* metadata version: Chef 0.90, mdadm 1.2

The metadata version has 3 notable impacts.

1. 0.90 is limited to 28 devices per RAID volume.
2. Max device size of 2TB.
3. Some older versions of grub and lilo only support RAID1 md devices with
metadata 0.90 as a boot device.