Load LaunchAgents as console user, adding plist and session_type options
Proposed changes to chef mac osx service provider. This adds two
resource parameters, @plist and @session_type and changes logic to Load
LaunchAgents as Console user

@plist: Adds the logic to handle
(https://github.com/chef/chef/issues/2200) by Giving the user the
option to pass a plist, in the case that the plist name and label name
don't match.

@session_type to help account launching
the service as the console user with a different session type.

( Im a Facebook employee and PhilD will do the final merge. )