Revert "Macports provider - provide package"
This reverts commit c793d2d5a4fe69c3e636239145139c0afb02fec1.

Fixes #3022.

Homebrew became the default package provider in Chef 12 via #1709, but some
users wanted to keep macports as their default provider so they didn't have
to update all their recipes for Chef 12. #2722 attempted to allow you to easily
change the defaults, but that caused a regression due to which resource was
being loaded, which was filed as #3022.