Disable Cmdlet tests on old versions of powershell
cmdlet.rb uses stream redirection, which is not available in
powershell 2. Since this is a helper class for our `dsc_script`
and `dsc_resource`, this is a fairly safe change.

While the verbose stream redirection is currently unused, it
will be useful if Invoke-DscResource logs its verbose output
to the verbose stream.