Allow tags to be set on a node during bootstrap
I can set attributes, environment, and run_list, but why not tags?

Some of our recipes are driven off of tags, such as for sudo access or
membership in a load balancing pool.

If this patch is accepted then I will add the corresponding feature to
knife-vsphere.

This patch only works with the new vault method, not the validator key
method. If someone has some pointers on where to add it in the validator
method I'd be happy to amend.