Allow knife sub-command loader to match platform specific gems.
This fixes issue #3272.
The underlying issue was that "-x86-mingw32"-platform suffixed chef gem
(as it is installed on windows) was filtered out by the regex.  Thus it
was detected as being "not from this chef version".

The PR adapts the regex and adds some valid / invalid specs for it.