windows_package is idempotent again
This was broken in #3034. The issue is that the package does not get detected
as installed because the version that is installed has \x00 at the end,
while the version from the msi does not. This fails comparison, and we fall
into code that does not use source and thus requires candidate version and we
die. (Or something like that)

Solves #3316