use same fix for upload as install metadata.json
metadata.json is a datafile, created by metadata.rb.
for whatever reason, the inclusion of one or the other is now a free choice.
This means that sometimes cookbooks may come with one and not the other.

Including the resulting DATA and not the CODE is now supported by prior changes when using the "knife cookbook site install"
However, that change was not incorporated globally and so we find that  "knife cookbook upload" continues to fail.

example failure:
+ echo 'Uploading consul to chef servers'
Uploading consul to chef servers
+ cd ..
+ which knife
/usr/bin/knife
+ knife -v
Chef: 12.1.2
+ knife cookbook upload consul -o .
Uploading consul         [0.6.0]
ERROR: Cookbook consul depends on cookbooks which are not currently
ERROR: being uploaded and cannot be found on the server.
ERROR: The missing cookbook(s) are: 'golang' version '~> 1.3.0'

The failure here is that the consul cookbook from supermarket does not come with metadata.rb, the code required to generate said metadata.json, and which is looked for by the "knife cookbook upload" command. A fix was put in place for "knife cookbook site install"

now, i can use knife cookbook metadata to generate a metadata.json, however there is no option for reverse engineering the metadata.rb from the data.

I've added the same logic to make upload work.
this supporting routine most likely belongs else where, however, for its true global reach.

I still question that it can be correct to exclude the CODE (metadata.rb) and include the DATA (metadata.json).
this feels broken to me with this PR being a hack on top of a bug to dress-up the bug.