change default log_location + log_level for knife
send logging to STDERR and drop default log_level to :warn
so that we can see warnings again.  logging needs to go to
STDERR so that we don't break someone's 'API' where they're piping
stdin to some filter.  we believe the log_level was not warn
in order to suppress warns off of 404s which no longer warn by
default now.