suppress cookbook self-dependencies and warn
cookbooks with self-deps:

name 'foo'
depends 'foo'

are useless and can waste cycles in the depsolver (particularly in
the non-solution case), and likely limit the possible choices of
depsolvers that we could pick at some point to replace gecode.

filtering out the self-dep here will prevent the chef server depsolver
from seeing the self-dep and needing to do the work to strip it.

also warn the user so that they can remove the self-dep.

in the future this will be a hard error.

foodcritic also has a rule about removing these.