Increase the Timeout::timeout in execute timeout test
Issue #2175 found that execute resource timeouts weren't working on Windows but
mixlib-shellout was waiting 600 seconds for the test to fail and still raising.
Timeout was added in #2686 to make sure the test failed in this case.

We're currently seeing intermittant failures on a FreeBSD tester, which is
likely a performance issue. But it's possible that we're seeing another issue
that we don't yet understand, so I don't want to throw away the regression
test yet.