Fix ability to monkey match LWRP through Chef::Resource::MyLwrp
Repro
=====
The following cookbook would repro the issue

cookbooks/lwrptest/resources/my_lwrp.rb
---------------------------------------
```ruby
```

cookbooks/lwrptest/providers/my_lwrp.rb
---------------------------------------
```ruby
```

cookbooks/lwrptest/recipes/default.rb
-------------------------------------
```ruby
module Something
  def something
    puts 'something here'
  end
end

::Chef::Resource::LwrptestMyLwrp.send(:include, Something)

lwrptest_my_lwrp 'blah' do
  something
end
```

Why it does not work
====================
Running this in 12.4.0 produces the error
```
NoMethodError
-------------
undefined method `something' for LWRP resource lwrptest_my_lwrp from cookbook lwrptest
```

The problem is that LWRP usage through `Chef::Resource::MyLwrp` and
`Chef::Provider::MyLwrp` is depreacted. In order to participate in
the deprecation, `Chef::Resource::MyLwrp` is a subclass of the actual
`MyLwrp` class with `initialize` overriden to do a Chef::Log.deprecation.

In the example above, the intention is to add a method to the resource
so that it can be used from the dsl. The reason this does not work is
because including the method on `Chef::Resource::MyLwrp` adds the method
to the subclass, and the resource that is looked up when running the
recipe is the actual resource.