Fix up specs
Loading the lwrp multiple times causes problems when comparing classes
due to the their dynamic nature. It worked fine when we were overriding
the things that checked.

I've left the tests mostly as is, other than the fact that the test
lwrp is loaded only once. The tests should still hold true, even
with the new implementation.