Make the doc formatter actually show what version of a cookbook is being used.
This has bugged me forever. It does change the API of the Formatter
system in a ~non-back-compat way but I don't think I've actually seen
any formatters outside of core Chef so how much do we want to worry
about this? We could write a shim in `Chef::EventDispatcher::Dispatcher`
to check the arity of the method if needed.