Changes --hide-healthy to --hide-by-mins MINS
Fixes #3679

Reasoning:

The definition of "healthy" is overloaded in the old form
(--hide-healthy) to be "a host that has run chef". The code
makes no check to determine if the chef run was successful,
so it has the capability to provide false positives for
"health". Just because a node object was saved with ohai_time
set doesn't mean the Chef run was successful. There are
exception handlers like lastrun that intentionally save the node
object on exception.

The previous 1 hour hardcoded time was totally arbitrary. Perhaps
"healthy" to others means "nodes that have run chef in the last
4 hours" (or 30 minutes, or 3 days...).