Undeprecate RunContext#resource_collection=.
I use this in Poise and would like to make sure it continues to be a
supported API. This is required because I use subcontexts with different
semantics from Chef's isolated ones. The alternative to un-deprecating
this would be to allow passing in an alternate factory other than
Chef::ResourceCollection.new().