fix errors when files go away during chown
The deploy resource attempts to change the owner of all files in the
deploy directory, but it breaks when files are removed during the
process. This is because FileUtils::chown_R gets a list of the files and
then walks through and updates them, and if that takes any time at  all
then files may be deleted during the process. When this happens, the
deploy will fail with Errno::ENOENT. chown_R takes an options hash with
:force as one of the available options, and the only thing it does is
stop errors from being raised under these circumstances.