openbsd rcctl service provider
Detect and use the rcctl(8) service management interface on OpenBSD
nodes if available.

This provider provides support for rcctl's flags mechanism by passing a
flags key to the service parameters attribute, eg:

```
service 'dhcpd' do
  parameters(flags: 'em1')
  action [:enable, :start]
end
```

Caveats:

* Not all rc.d scripts will support reload. We should detect this and
remove it from the list of :supported actions. As written, if you try to
reload a service that doesn't support it the underlying rc.d/rcctl calls
will fail and the resource will be fail to converge.

* Right now parameters are only checked and set on action_enable since
rcctl(8) only allows setting parameters on enabled services. It might
be desirable to also set parameters in action_start in the case that
action_start is processed before parameters are set and causes the
daemon startup to fail. Same for action_restart.