Use the proper python interpretor for yum-dump.py on Fedora 21+
Fedora 21+ use dnf as the primary package manager.  Lamont added code in
12.5 to allow for a yum compat mode. This doesn't entirely work though
as we need yum-dump.py to correctly run.  We were parsing the shabang in
the yum binary to find the path to python.  On a dnf system the yum
binary is a bash script though so we were trying to run yum-dump.py
using bash which obviously fails.  I added a fallback method to use
python if the shebang parsing returns bash. With this in place AND the
yum package installed you can use the package resource on Fedora.