fix couple of issues in dpkg cleanup
this is just code rearrangement, no changes to behavior at
all.  pushes the source_array and resolved_source_array up
into the superclass so that other providers can use it (as it
was intended, but that slipped out of the last PR).  also
renamed it to something a little longer and less likely to
clash with variables (which it was already doing).

the block comment also refers to an earlier iteration that used
coercions, which had to be backed out because of chefspec.