restore handling of dpkg -s returns codes (w/tests)
Previously dpkg -s would allow a 0 or 1 exit status and it wasn't
entirely clear why this was so either from the code or the tests.

This restores throwing an exception if we are outside of the [0,1]
range, and then adds tests for both behaviors:

- on old ubuntu/debian we get an exit(0) and output on stdout if the
  file is not installed
- on newer ubuntu/debian we get an exit(1) and output on stderr if
  the file is not installed

Added tests for both those cases, and the case where dpkg -s does
some other kind of exitcode barf and we should raise.