Fix dpkg specs to run on debian 6
The exit codes on debian 6 do not match what was expected.
Purging seems to behave differently... metadata is not purged.
We now check the contents of the output to see if the output
has been removed or purged. There are also places where we
dont care if it should be purged or removed... either would
suffice. This case is also represented, and all checks
go through `should_be_purged_or_removed`.