tags always an array; fix set_unless
previously if the node had a { normal: { tags: nil } } that we read
since the key existed (but was nil) we would not initialize with an
array.  replacing the code with a call to node.set_unless revealed that
set_unless was similarly buggy.  fixed both issues by fixing set_unless.