Included the message part to the body of :delete request
While RFC does not explicitly suggest that, there are services like CloudFlare that utilize the data part of DELETE request as well. It actually does not execute the requests with out it making Chef http_request unsuitable for executing actions like purge requests. I added the message part to the body of :delete.