fix Lint/DuplicatedKey
so, this is a bit special...

the API for "priority" is clearly broken.  but this broken API comes
from the service resource itself which appears to have been broken
by design since at least 2010.

just fixing the code for how it behaves and dropping in FIXMEs
for now, since this uncovered a fairly massive API-breaking yak.