Don't check the key for auth_credentials_spec
The AuthCredentials class does not do anything special
with regards to the key. We should only need to check
that the one we passed in, not that it's pem representation.

The reason for removing this extra (unneeded) check is
because it breaks when openssl is in fips mode. When in
that mode, openssl with generate a PKCS#8 pem representation, while
the thing that we checked against was a PKCS#1 pem representation.