preferred_manifest_record: fix pretty print.
The `preferred_manifest_record` function pretty prints the list of
existing files when the asked file can't be found. To do that, the root
path of the cookbook is removed from each existing file but the update
is done on the `existing_files` variable, not on a copy. When called
several times, this variable will eventually be equals to `[nil]`,
leading to a NoMethodError.

Without the fix, the new unit test fails with

> expected Chef::Exceptions::FileNotFound, got #<NoMethodError: undefined method `[]' for nil:NilClass> with backtrace:
>   # ./lib/chef/cookbook_version.rb:321:in `block in preferred_manifest_record'
>   # ./lib/chef/cookbook_version.rb:319:in `map!'
>   # ./lib/chef/cookbook_version.rb:319:in `preferred_manifest_record'
>   # ./spec/unit/cookbook_version_file_specificity_spec.rb:318:in `block (3 levels) in <top (required)>'
>   # ./spec/unit/cookbook_version_file_specificity_spec.rb:317:in `block (2 levels) in <top (required)>'

Fixes #2561.