wire up chef handlers directly from libraries
smear just a little bit of minimal DSL wiring into the Chef::Report
class so that classes can wire themselves up automatically as
handlers.

this also extends the Chef::Config[*_handler] arrays to take classes
as well as instances.

in the case of passing a class that is not a singleton, a new instance
will be created when the report is run.  it will obviously be difficult
to pass any state into this instance.

if the class that is passed supports calling `#instance` we assume it
is a Singleton and get the instance and use that.  this is to support
other chef code grabbing/creating the singleton instance and poking
it with data (probably configuring where the reporter goes and what
creds it needs to publish and other stuff).