Adding node object to run_converge message, include ignore_failure property
Many use cases that involve consuming the run_converge messages and
displaying them to end users include needing additional data about
the node that generated the run_converge message. This change
consolidates the run_converge and node_update message into a single
message making it easier for users to filter run_converge messages
based on attributes and criteria of the node itself.

Additionally, the ignore_failure property on a resource is now
included for each resource in the resource list so end users can
decided whether the failure was a hard failure or a soft failure.