In Travis use latest ruby releases + trusty
The nice rvm aliases in Travis (2.1, 2.2, etc) have never been updated
and point to some really old releases. We should specify the latest
releases of these feature releases. Also if we require sudo then we're
running the GCE environment and we can also specify that we want to run
on the Trusty "beta" image which has much newer deps all around.

Signed-off-by: Tim Smith <tsmith@chef.io>