launchd: start_calendar_interval validity checking
StartCalendarInterval has some gotchas so we coerce it to sanity check.

According to `man 5 launchd.plist`:
  StartCalendarInterval [is a] <dictionary of integers or array of dictionaries of integers>
    ... Missing arguments are considered to be wildcard.

What the man page doesn't state, but what was observed (OSX 10.11.5, launchctrl v3.4.0) is that key that are specified, _but with invalid values, will also be silently treated as a wildcard values.

This means that an entry like:
  { "Hour"=>0, "Weekday"=>"6-7"}
will not just run on midnight of Sat and Sun, rather it will run _every_ midnight.

We check to make sure that the values are integers and not strings, adhering to the manpage documentation.

This also ensures that only keys that StartCalendarInterval supports will be passed.
This eliminates passing a key (like `:hour` or `'hour') and thinking it works when in reality launchd is just treating it as a wildcard.

NOTE: This is a "breaking" change in that: it will break existing cookbooks that are using the current docs: https://docs.chef.io/resource_launchd.html (See start_calendar_interval); however the generated plist doesn't appear to actually do what was intended anyway.